virtqueue: Abstract shared memory handling
Abstract the handling of the shared memory region to allow for more
optimized implementation based on system specific considerations about
how memory is dealt with.

We treat the sh_mem region in the virtqueue as an obaque pointer that
only should get used by v2p and p2v functions.  The pointer will provide
whatever data the implmenetation needs to allow for the translation in
those functions to happen.

We mark virtqueue_phys_to_virt and virtqueue_virt_to_phys as weak to
allow other implementation.  By default we provide a solution around
libmetal and metal_io_region.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>